+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Refactored API Output to an Abstract Class so that it is easy for extension developers to develop their own implementation.
 * Added Markup/Html Field Value output field handler.
 * Improved In Memory Dataflow so that joins and filters would work.
+* Improved Contact Summary Tab output so it includes a count.
 
 # Version 1.6.0